Slice-ID: Clean up a little.
It's hackathon code, but we won't have that much time afterwards to make it production-ready, so clean up _just_ a little before things become permanent.

part of UMH-2025
