+++ CMakeLists.txt
@@ -141,6 +141,7 @@ set(engine_SRCS # Except main.cpp.
         src/utils/ExtrusionLine.cpp
         src/utils/ExtrusionSegment.cpp
         src/utils/gettime.cpp
+        src/utils/IdFieldInfo.cpp
         src/utils/LabelMaker.cpp
         src/utils/linearAlg2D.cpp
         src/utils/ListPolyIt.cpp
@@ -168,8 +169,6 @@ set(engine_SRCS # Except main.cpp.
         src/utils/scoring/PositionBasedScoringCriterion.cpp
         src/utils/scoring/RandomScoringCriterion.cpp
         src/utils/scoring/TextureScoringCriterion.cpp
-
-        src/utils/types/idfieldinfo.cpp
 
         src/geometry/Point2LL.cpp
         src/geometry/Point3LL.cpp
