+++ src/mesh.cpp
@@ -3,16 +3,48 @@
 
 #include "mesh.h"
 
-#include <algorithm>
 #include <numbers>
 
 #include <spdlog/spdlog.h>
 
 #include "utils/Point3D.h"
-#include "utils/types/idfieldinfo.h"
 
 namespace cura
 {
+
+void Image::visitSpanPerPixel(const Point2F& a, const Point2F& b, const std::function<void(const int32_t, const Point2F&)>& func) const
+{
+    constexpr auto func_major_stepper =
+        [](const int64_t& da, const int64_t& abs_db, const int64_t& i)
+        {
+            return da < 0 ? -i : i;
+        };
+    constexpr auto func_minor_stepper =
+        [](const int64_t& da, const int64_t& abs_db, const int64_t& i)
+        {
+            return (i * da) / abs_db;
+        };
+
+    const auto x0 = static_cast<int64_t>(a.x_ * width_);
+    const auto y0 = static_cast<int64_t>(a.y_ * height_);
+    const auto x1 = static_cast<int64_t>(b.x_ * width_);
+    const auto y1 = static_cast<int64_t>(b.y_ * height_);
+
+    const auto dx = x1 - x0;
+    const auto dy = y1 - y0;
+    const auto abs_dx = std::llabs(dx);
+    const auto abs_dy = std::llabs(dy);
+    const auto max_span = std::max(abs_dx, abs_dy);
+
+    const auto func_x = abs_dy <= abs_dx ? func_major_stepper : func_minor_stepper;
+    const auto func_y = abs_dx <= abs_dy ? func_major_stepper : func_minor_stepper;
+    for (size_t i_pix = 0; i_pix <= max_span; i_pix++)
+    {
+        const auto xi = x0 + func_x(dx, abs_dy, i_pix);
+        const auto yi = y0 + func_y(dy, abs_dx, i_pix);
+        func(getPixel(xi, yi), Point2F(static_cast<float>(xi) / width_, static_cast<float>(yi) / height_));
+    }
+}
 
 const int vertex_meld_distance = MM2INT(0.03);
 /*!
